requirements.txt
cupy==7.8.0
matplotlib==3.3.2
xgboost==1.2.1
rmm==0.16.0
matplotlib_venn==0.11.5
xfeat==0.1.1
cuml==0.16.0
pandas==1.1.3
seaborn==0.11.0
torch==1.6.0
catboost==0.24.2
tqdm==4.50.2
requests==2.24.0
numpy==1.19.2
scipy==1.5.3
imbalanced_learn==0.7.0
cudf==0.16.0
imblearn==0.0
KTBoost==0.1.19
pytorch_tabnet==2.0.1
PyYAML==5.3.1
scikit_learn==0.23.2
